requirements-dev.txt
# Requirements for development environment
# Install with: pip install -r requirements-dev.txt

-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
flake8==6.1.0
black==23.12.1
isort==5.13.2

# Debugging
django-debug-toolbar==4.2.0
ipython==8.18.1
